Code refactoring
-refactored code to simply the main code
-separated the cdc-acm serial code and tasks
-separated functions the control the soc cores
