+++ Quark_Sketch/src/Makefile
@@ -1,4 +1,4 @@
-obj-y += blink.o arduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o sharedmemory_com.o
-#obj-y += blank.o arduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o sharedmemory_com.o
-#obj-y += smc.o arduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o sharedmemory_com.o
-obj-y += pinInterrupts.o arduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o sharedmemory_com.o
+obj-y += blink.o sharedmemory_com.o soc_ctrl.o cdcacm_serial.o arduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o
+#obj-y += blank.o sharedmemory_com.o soc_ctrl.o cdcacm_serial.o arduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o
+#obj-y += smc.o sharedmemory_com.o soc_ctrl.o cdcacm_serial.oarduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o
+#obj-y += pinInterrupts.o sharedmemory_com.o soc_ctrl.o cdcacm_serial.o arduino/wiring_digital.o arduino/wiring_analog.o arduino/variant.o arduino/WInterrupts.o arduino/wiring.o arduino/serial_print.o
